% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-zouwquantreg.R
\name{methods-zouwquantreg}
\alias{methods-zouwquantreg}
\alias{print.zouwquantreg}
\alias{summary.zouwquantreg}
\alias{coef.zouwquantreg}
\alias{vcov.zouwquantreg}
\alias{logLik.zouwquantreg}
\alias{confint.zouwquantreg}
\alias{fitted.zouwquantreg}
\title{Methods for \code{zouwquantreg} Objects}
\usage{
\method{print}{zouwquantreg}(x, digits = 4, ...)

\method{summary}{zouwquantreg}(object, ...)

\method{coef}{zouwquantreg}(object, ...)

\method{vcov}{zouwquantreg}(object, ...)

\method{logLik}{zouwquantreg}(object, ...)

\method{confint}{zouwquantreg}(object, parm, level = 0.95, what = c("both", "quant", "infl"), ...)

\method{fitted}{zouwquantreg}(object, type = c("mu", "phi", "nu", "all"), ...)
}
\arguments{
\item{digits}{minimal number of _significant_ digits
the estimated parameters is returned and printed. Default is \code{FALSE}.}

\item{...}{additional argument(s) for methods. Currently not used.}

\item{object, x}{fitted model object of class \code{\link{zouwquantreg}}.}

\item{parm}{a specification of which parameters are to be given confidence intervals,
either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{what}{character indicating what component return the confidence intervals.
If \code{quant} return coefficients from quantile model,
if \code{infl} return coefficients from inflation model. If \code{both} a list with parameters from
quantile and inflation models.}

\item{type}{character indicating type of fitted values to return.}
}
\description{
Methods for extracting information from fitted unit-Weibull quantile regression
objects of class \code{\link{zouwquantreg}}.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
