% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.uwquantreg}
\alias{predict.uwquantreg}
\title{Prediction Method for \code{uwquantreg} Objects}
\usage{
\method{predict}{uwquantreg}(
  object,
  newdata,
  type,
  interval = "confidence",
  level = 0.95,
  se.fit = FALSE,
  ...
)
}
\arguments{
\item{object}{fitted model object of class \code{\link{uwquantreg}}.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict.
If omitted, the original observations are used.}

\item{type}{character indicating type of predictions. The options are \code{mu}, \code{phi} and \code{all}.}

\item{interval}{type of interval desired. The options are \code{none} and \code{confidence}.}

\item{level}{coverage probability for the confidence intervals. Default is \code{0.95}.}

\item{se.fit}{logical. If \code{TRUE} return the asymptotic standard errors.}

\item{...}{currently not used.}
}
\description{
Extract various types of predictions from unit-Weibull quantile regression models.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
