% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.zouwquantreg}
\alias{predict.zouwquantreg}
\title{Prediction Method for \code{zouwquantreg} Objects}
\usage{
\method{predict}{zouwquantreg}(object, newdata, type = "quantile", ...)
}
\arguments{
\item{object}{fitted model object of class \code{\link{uwquantreg}}.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict.
If omitted, the original observations are used.}

\item{type}{character indicating type of predictions. The currently option is \code{quantile}.}

\item{...}{currently not used.}
}
\description{
Extract various types of predictions from Zero-or-One unit-Weibull quantile regression models.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
