% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.zouwquantreg}
\alias{residuals.zouwquantreg}
\title{Residuals Method for \code{zouwquantreg} Objects}
\usage{
\method{residuals}{zouwquantreg}(object, ...)
}
\arguments{
\item{object}{fitted model object of class \code{\link{zouwquantreg}}.}

\item{...}{currently not used.}
}
\description{
Extract quantile residuals from Zero-or-One unit-Weibull quantile regression models.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
