% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{transport}
\alias{transport}
\title{Transport in Campus}
\format{
A \code{\link{data.frame}} with 60 observations and 7
  columns:

\itemize{
\item \code{ntrips }: number of trips to campus in the past four weeks.
\item \code{nbiked  }: the number of times the respondent biked to campus.
\item \code{status }: respondents status (student/faculty/staff);
\item \code{gender }: person's gender.
\item \code{parking }: duration of parking permit (6, 9, or 12-month.
\item \code{distance }: distance to campus.
\item \code{propbiked }: proportion of time a respondent biked to campus.
}
}
\usage{
data(transport, package = "uwquantreg")
}
\description{
This data set is regarding the mode of transportation in a campus. A stratified sample of
60 respondents was drawn for the purpose of oversampling people who sometimes bike to campus.
}
\references{
Korosteleva, O., 2019. \bold{Advanced Regression Models with SAS and R}. Taylor & Francis Group.
}
\keyword{datasets}
