% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trees}
\alias{trees}
\title{Mortality of Young Trees}
\format{
A \code{\link{data.frame}} with 26 observations and 7
  columns:

\itemize{
\item \code{planted }: number of planted trees.
\item \code{survived  }: number of trees that survived for two years.
\item \code{pest }: frequency of pest control;
\item \code{fertilization }: frequency of soil fertilization.
\item \code{precip }: average annual precipitation (in inches).
\item \code{wind }: average annual wind speed (in miles per hour).
\item \code{prop }: proportion of survived trees.
}
}
\usage{
data(trees, package = "uwquantreg")
}
\description{
Study conducting by Parks and Recreation Department on mortality of young trees planted in parks.
}
\references{
Korosteleva, O., 2019. \bold{Advanced Regression Models with SAS and R}. Taylor & Francis Group.
}
\keyword{datasets}
