% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr.R
\name{uweibull}
\alias{uweibull}
\alias{duweibull}
\alias{puweibull}
\alias{quweibull}
\alias{ruweibull}
\title{unit-Weibull distribution}
\usage{
duweibull(x, mu, phi, tau, log = FALSE)

puweibull(q, mu, phi, tau, lower.tail = TRUE, log.p = FALSE)

quweibull(p, mu, phi, tau, lower.tail = TRUE, log.p = FALSE)

ruweibull(n, mu, phi, tau)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{location parameter indicating the \eqn{\tau}-quantile \eqn{\tau \in (0, 1)}.}

\item{phi}{nonnegative shape parameter.}

\item{tau}{the parameter to specify which quantile use in the parametrization.}

\item{log, log.p}{logical; If TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; If TRUE, (default), \eqn{P(X \leq x)} are returned, otherwise \eqn{P(X > x)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{duweibull} gives the density, \code{puweibull} gives the distribution function,
\code{quweibull} gives the quantile function and \code{ruweibull} generates random deviates.

Invalid arguments will return an error message.
}
\description{
Density function, distribution function, quantile function, random number generation function
for the unit-Weibull distribution re-parametrized in terms of the \eqn{\tau}th quantile.
}
\examples{
set.seed(6969)
x <- ruweibull(n = 1000, mu = 0.5, phi = 1.5, tau = 0.5)
R <- range(x)
S <- seq(from = R[1], to = R[2], by = 0.01)
hist(x, prob = TRUE, main = 'unit-Weibull distribution')
lines(S, duweibull(x = S, mu = 0.5, phi = 1.5, tau = 0.5), col = 2)
plot(ecdf(x))
lines(S, puweibull(q = S, mu = 0.5, phi = 1.5, tau = 0.5), col = 2)
plot(quantile(x, probs = S), type = "l")
lines(quweibull(p = S, mu = 0.5, phi = 1.5, tau = 0.5), col = 2)
}
\references{
Mazucheli, J., Menezes, A. F. B and Ghitany, M. E. (2018). The unit-Weibull distribution and
associated inference. \emph{Journal of Applied Probability and Statistics} \bold{13} (2), 1--22.

Mazucheli, J., Menezes, A. F. B., Fernandes, L. B., Oliveira, R. P., Ghitany, M. E., 2020.
The unit-Weibull distribution as an alternative to the Kumaraswamy distribution for the
modelling of quantiles conditional on covariates. \emph{Jounal of Applied Statistics} \bold{47} (6), 954--974.
}
\author{
André Felipe B. Menezes \email{andrefelipemaringa@gmail.com}
}
