% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwquantreg.R
\name{uwquantreg.control}
\alias{uwquantreg.control}
\title{Control Parameters for unit-Weibull quantile regression}
\usage{
uwquantreg.control(method = "BFGS", maxit = 5000, reltol = 1e-08, ...)
}
\arguments{
\item{method}{characters string specifying the method argument passed to \code{optim}.}

\item{maxit}{integer specifying the \code{maxit} argument (maximal number of iterations) passed to \code{optim}.}

\item{reltol}{relative convergence tolerance passed to \code{optim}.}

\item{...}{arguments passed to \code{optim}.}
}
\description{
Parameters that control fitting of unit-weibull quantile regression models using \code{uwquantreg}.
}
\references{
Mazucheli, J., Menezes, A. F. B., Fernandes, L. B., Oliveira, R. P., Ghitany, M. E., 2020.
The unit-Weibull distribution as an alternative to the Kumaraswamy distribution for the
modelling of quantiles conditional on covariates. \emph{Jounal of Applied Statistics} \bold{47} (6), 954--974.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
