% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwquantreg.R
\name{uwquantreg}
\alias{uwquantreg}
\alias{uwquantreg.fit}
\title{unit-Weibull quantile regression}
\usage{
uwquantreg(
  formula,
  tau,
  data,
  link,
  link.phi = NULL,
  start = NULL,
  control = uwquantreg.control()
)

uwquantreg.fit(
  y,
  X,
  Z = NULL,
  tau,
  link,
  link.phi,
  start,
  control = uwquantreg.control()
)
}
\arguments{
\item{formula}{symbolic description of the quantile model.}

\item{tau}{the quantile(s) to be estimated, number between 0 and 1.}

\item{data}{data.frame contain the variables in the model.}

\item{link}{character specification of the link function in the quantile model. Default is \code{logit}.}

\item{link.phi}{character specification of the link function in the shape model. Default is \code{log}.}

\item{start}{an optional vector with starting values for all parameters}

\item{control}{a list of control arguments specified via \code{uwquantreg.control}.}

\item{y}{numeric vector of response variable.}

\item{X}{numeric matrix. Regressor matrix for the quantile model.}

\item{Z}{numeric matrix. Regressor matrix for the shape model. Default is constant shape model.}
}
\description{
Fit unit-Weibull quantile regression.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
