% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr.R
\name{zouweibull}
\alias{zouweibull}
\alias{dzouweibull}
\alias{pzouweibull}
\alias{qzouweibull}
\alias{rzouweibull}
\title{Zero-or-One unit-Weibull distribution}
\usage{
dzouweibull(x, mu, phi, tau, nu, inflation, log = FALSE)

pzouweibull(q, mu, phi, tau, nu, inflation, lower.tail = TRUE, log.p = FALSE)

qzouweibull(p, mu, phi, tau, nu, inflation, lower.tail = TRUE, log.p = FALSE)

rzouweibull(n, mu, phi, tau, nu, inflation)
}
\arguments{
\item{x, q}{vector of positive quantiles.}

\item{mu}{location parameter indicating the \eqn{tau}-quantile \eqn{\tau \in (0, 1)}.}

\item{phi}{nonnegative shape parameter.}

\item{tau}{the parameter to specify which quantile use in the parametrization.}

\item{nu}{probability of inflation.}

\item{inflation}{specify the zero (0) or the one (1) inflation.}

\item{log, log.p}{logical; If TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; If TRUE, (default), \eqn{P(X \leq x)} are returned, otherwise \eqn{P(X > x)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{dzouweibull} gives the density, \code{pzouweibull} gives the distribution function,
\code{qzouweibull} gives the quantile function and \code{rzouweibull} generates random deviates.

Invalid arguments will return an error message.
}
\description{
Density function, distribution function, quantile function, random number generation function
for the zero-or-one unit-Weibull distribution re-parametrized in terms of the \eqn{\tau}th quantile.
}
\examples{
set.seed(6969)
x <- rzouweibull(n = 1e4, mu = 0.5, phi = 1.5, tau = 0.5, nu = 0.3, inflation = 1)
R <- range(x)
S <- seq(from = R[1], to = R[2], l = 1000)
plot(ecdf(x))
lines(S, pzouweibull(q = S, mu = 0.5, phi = 1.5, tau = 0.5, nu = 0.3, inflation = 1), col = 2)
plot(quantile(x, probs = S), type = "l")
lines(qzouweibull(p = S, mu = 0.5, phi = 1.5, tau = 0.5, nu = 0.3, inflation = 1), col = 2)

probs <- c(0.1, 0.2, 0.5, 0.7, 0.9)
(p <- pzouweibull(q = probs, mu = 0.1, phi = 2, tau = 0.5, nu = 0.5, inflation = 1))
qzouweibull(p, mu = 0.1, phi = 2, tau = 0.5, nu = 0.5, inflation = 1)

(p <- pzouweibull(q = probs, mu = 0.1, phi = 2, tau = 0.5, nu = 0.5, inflation = 0))
qzouweibull(p, mu = 0.1, phi = 2, tau = 0.5, nu = 0.5, inflation = 0)

}
\references{
Mazucheli, J., Menezes, A. F. B and Ghitany, M. E. (2018). The unit-Weibull distribution and
associated inference. \emph{Journal of Applied Probability and Statistics} \bold{13} (2), 1--22.

Mazucheli, J., Menezes, A. F. B., Fernandes, L. B., Oliveira, R. P., Ghitany, M. E., 2020.
The unit-Weibull distribution as an alternative to the Kumaraswamy distribution for the
modelling of quantiles conditional on covariates. \emph{Jounal of Applied Statistics} \bold{47} (6), 954--974.
}
\seealso{
\code{\link{uweibull}}.
}
\author{
André Felipe B. Menezes \email{andrefelipemaringa@gmail.com}
}
