% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zouwquantreg.R
\name{zouwquantreg}
\alias{zouwquantreg}
\title{Zero-or-One unit-Weibull quantile regression}
\usage{
zouwquantreg(
  formula,
  tau,
  inflation,
  data,
  link,
  link.phi = NULL,
  formula.nu,
  link.nu,
  start = NULL,
  control = uwquantreg.control()
)
}
\arguments{
\item{formula}{symbolic description of the quantile model.}

\item{tau}{the quantile(s) to be estimated, number between 0 and 1.}

\item{inflation}{specify the zero (0) or the one (1) inflation.}

\item{data}{data.frame contain the variables in the model.}

\item{link}{character specification of the link function in the quantile model. Default is \code{logit}.}

\item{link.phi}{character specification of the link function in the shape model. Default is \code{log}.}

\item{formula.nu}{a symbolic description of the model to be fitted for the inflation part.}

\item{link.nu}{character specification of the link function inflation model. Default is \code{logit}.}

\item{start}{a list with length equal specifying starting values for the parameters in the linear predictor.
The first entrie refers to continous part and the second for the inflation. Default is \code{NULL}}

\item{control}{a list of control arguments specified via \code{uwquantreg.control}.}
}
\description{
Fit Zero-or-One unit-Weibull quantile regression.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
